Refactor notification related frontend code
